chore(EncryptionContactChip): Put generatePassword back in viewModel
