+++ app/src/main/java/com/infomaniak/mail/ui/newMessage/encryption/EncryptionPasswordFragment.kt
@@ -35,7 +35,6 @@ import com.infomaniak.mail.ui.newMessage.ContactChipAdapter
 import com.infomaniak.mail.ui.newMessage.NewMessageViewModel
 import com.infomaniak.mail.utils.extensions.copyStringToClipboard
 import dagger.hilt.android.AndroidEntryPoint
-import java.security.SecureRandom
 import javax.inject.Inject
 
 @AndroidEntryPoint
@@ -73,12 +72,12 @@ class EncryptionPasswordFragment : Fragment() {
 
     private fun setupPasswordTextField() = with(binding) {
         passwordInputLayout.setEndIconOnClickListener {
-            passwordInput.setText(generatePassword())
+            passwordInput.setText(encryptionViewModel.generatePassword())
         }
         passwordInput.apply {
             doOnTextChanged { password, _, _, _ -> newMessageViewModel.encryptionPassword.value = password.toString() }
             val initialPassword = newMessageViewModel.encryptionPassword.value.takeUnless { it.isNullOrBlank() }
-            setText(initialPassword ?: generatePassword())
+            setText(initialPassword ?: encryptionViewModel.generatePassword())
         }
     }
 
@@ -92,21 +91,7 @@ class EncryptionPasswordFragment : Fragment() {
         }
     }
 
-    private fun generatePassword(): String {
-        val generator = SecureRandom.getInstanceStrong()
-        var generatedPassword = ""
-        val charactersSetCount = PASSWORD_CHARACTERS_SET.count()
-        (0..<PASSWORD_MIN_LENGTH).forEach {
-            generatedPassword += PASSWORD_CHARACTERS_SET[generator.nextInt(charactersSetCount)]
-        }
-
-        return generatedPassword
-    }
-
     companion object {
         private const val ENCRYPTION_FAQ_URL = "https://faq.infomaniak.com/1582"
-        private const val PASSWORD_MIN_LENGTH = 16
-        private const val PASSWORD_CHARACTERS_SET =
-            "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()-_=+[]{}|;:,.<>?"
     }
 }
