+++ app/src/main/java/com/infomaniak/mail/ui/newMessage/encryption/EncryptionViewModel.kt
@@ -32,6 +32,7 @@ import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
+import java.security.SecureRandom
 import javax.inject.Inject
 
 @HiltViewModel
@@ -88,10 +89,27 @@ class EncryptionViewModel @Inject constructor(
         }
     }
 
+    fun generatePassword(): String {
+        val generator = SecureRandom.getInstanceStrong()
+        var generatedPassword = ""
+        val charactersSetCount = PASSWORD_CHARACTERS_SET.count()
+        (0..<PASSWORD_MIN_LENGTH).forEach { _ ->
+            generatedPassword += PASSWORD_CHARACTERS_SET[generator.nextInt(charactersSetCount)]
+        }
+
+        return generatedPassword
+    }
+
     private fun clearDataAndPostResult(recipients: Set<String>) {
         emailsBeingChecked.clear()
         unencryptableRecipients.postValue(recipients)
     }
 
     private class AutoBulkCallCancellationException : CancellationException()
+
+    companion object {
+        private const val PASSWORD_MIN_LENGTH = 16
+        private const val PASSWORD_CHARACTERS_SET =
+            "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*()-_=+[]{}|;:,.<>?"
+    }
 }
